// frontend/src/app/shared/api/model/create-chat-message-request.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateChatMessageRequest { 
  [key: string]: any | any;


    message: string;
    /**
     * Unikalny identyfikator
     */
    idImage?: number;
}