// frontend/src/app/shared/api/model/list-chat-messages200-response.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ListChatMessages200ResponseResultsInner } from './list-chat-messages200-response-results-inner';


export interface ListChatMessages200Response { 
  [key: string]: any | any;


    count: number;
    pages: number;
    currentPage: number;
    results: Array<ListChatMessages200ResponseResultsInner>;
}