// frontend/src/app/shared/api/model/list-chat-messages200-response-results-inner.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ListChatMessages200ResponseResultsInnerUser } from './list-chat-messages200-response-results-inner-user';
import { GetMeRestaurant200ResponseImage } from './get-me-restaurant200-response-image';
import { MessageType } from './message-type';


export interface ListChatMessages200ResponseResultsInner { 
  [key: string]: any | any;


    /**
     * Unikalny identyfikator
     */
    idMessage: number;
    user: ListChatMessages200ResponseResultsInnerUser;
    message: string;
    image: GetMeRestaurant200ResponseImage;
    type: MessageType;
    created: string;
}
export namespace ListChatMessages200ResponseResultsInner {
}