// frontend/src/app/shared/api/model/message-type.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type MessageType = 'Status' | 'Zgłoszenie' | 'Wiadomość';

export const MessageType = {
    Status: 'Status' as MessageType,
    Zgoszenie: 'Zgłoszenie' as MessageType,
    Wiadomo: 'Wiadomość' as MessageType
};